added : store employee frontend functionality WM2-5 WM2-8

+++ src/main/java/ma/yc/service/CandidatureService.java
@@ -13,7 +13,7 @@ public interface CandidatureService {
 
     boolean delete ( Candidature employee );
 
-    Candidature findById ( int id );
+    Candidature findById ( Long id );
 
     List<Candidature> findAll ();
 }
